Organize bundled WASM tests
